Modified shared array to unsigned long

+++ server.c
@@ -8,8 +8,8 @@ int main(int argc, char** argv[]){
         return(0);
     }
 
-    int* sh_ptr = create_shared_ptr(key);
-    if(sh_ptr == (int *)NULL){
+    unsigned long* sh_ptr = create_shared_ptr(key);
+    if(sh_ptr == (unsigned long*)NULL){
         return(0);
     }
